Fixed bad getInput + cleaned up tests with helpers
The tokenizer still fails - hangs upon mount
