+++ test/helpers/getInput.ts
@@ -1,0 +1,14 @@
+import { ReactWrapper } from 'enzyme';
+import { Props as TProps } from '../../src/Typeahead';
+import { Option } from '../../src/types';
+
+export default (component: ReactWrapper<TProps<Option>>) => {
+  let controlComponent = component.findWhere(
+    n => n.type() === 'input' && n.props().type !== 'hidden'
+  );
+  if (controlComponent.length === 1) {
+    return controlComponent.first();
+  }
+
+  throw new Error('No input found :-(');
+};
