+++ test/react-typeahead.test.tsx
@@ -1,26 +1,18 @@
 import * as React from 'react';
-import { renderIntoDocument, isCompositeComponent } from 'react-dom/test-utils';
+import { shallow } from 'enzyme';
 import {
   Typeahead as ReactTypeahead,
   Tokenizer as ReactTokenizer,
 } from '../src';
 
 describe('Main entry point', () => {
-  test('exports a Typeahead component', () => {
-    const typeahead = renderIntoDocument(
-      <ReactTypeahead options={['a', 'b']} />
-    );
-    // @ts-ignore
-    if (!typeahead) throw new Error('Typeahead not created!');
-    expect(isCompositeComponent(typeahead)).toBeTruthy();
+  test('exports a Typeahead component with an input', () => {
+    const typeahead = shallow(<ReactTypeahead options={['a', 'b']} />);
+    expect(typeahead.find('input')).toHaveLength(1);
   });
 
-  test('exports a Tokenizer component', () => {
-    const tokenizer = renderIntoDocument(
-      <ReactTokenizer options={['a', 'b']} />
-    );
-    // @ts-ignore
-    if (!tokenizer) throw new Error('Tokenizer not created!');
-    expect(isCompositeComponent(tokenizer)).toBeTruthy();
+  test('exports a Tokenizer component that contains a Typeahead', () => {
+    const tokenizer = shallow(<ReactTokenizer options={['a', 'b']} />);
+    expect(tokenizer.find('Typeahead')).toHaveLength(1);
   });
 });
