+++ test/setupTest.ts
@@ -5,8 +5,6 @@
  * @copyright 2017 Airbnb, Inc.
  */
 import enzyme from 'enzyme';
-// tslint:disable-next-line:variable-name
-// @ts-ignore
 import Adapter from 'enzyme-adapter-react-16';
 
 enzyme.configure({ adapter: new Adapter() });
